Genes now generated using seed
Seed is generated and set as a variable in Biomorph. No longer passed
into Gene as this would reset the seed every time a new Gene is created.

Factory class gene generation methods no longer take the Seed as a
parameter. The gene values are calculated straight after setting the
seed in Biomorph.

Separate constructor for Genes that are branches.

Cleaned up some code in Biomorph class.

+++ EvolutionaryArtStage2/src/main/BioGeneration.java
@@ -19,7 +19,7 @@ public class BioGeneration {
 	private Gene g1;
 	private Gene g2;
 	
-	private int seed; 
+	private long seed; 
 	private String bioData; //bioData Structure: {SEED; PARENT; CHILD1; CHILD2; CHILD3; CHILD4; CHILD5, CHILD6; CHILD7; CHILD8; CHILD9;}
 	
 	public BioGeneration(Biomorph b){
